Consolidating Schelling example
