+++ examples/Schelling/schelling.py
@@ -2,12 +2,39 @@ from __future__ import division  # For Python 2.x compatibility
 
 import random
 
-from mesa import Model
+from mesa import Model, Agent
 from mesa.time import RandomActivation
 from mesa.space import SingleGrid
 from mesa.datacollection import DataCollector
 
-from model.agents import SchellingAgent
+class SchellingAgent(Agent):
+    '''
+    Schelling segregation agent
+    '''
+    def __init__(self, pos, agent_type):
+        '''
+         Create a new Schelling agent.
+
+         Args:
+            unique_id: Unique identifier for the agent.
+            x, y: Agent initial location.
+            agent_type: Indicator for the agent's type (minority=1, majority=0)
+        '''
+        self.unique_id = pos
+        self.pos = pos
+        self.type = agent_type
+
+    def step(self, model):
+        similar = 0
+        for neighbor in model.grid.neighbor_iter(self.pos):
+            if neighbor.type == self.type:
+                similar += 1
+
+        # If unhappy, move:
+        if similar < model.homophily:
+            model.grid.move_to_empty(self)
+        else:
+            model.happy += 1
 
 
 class SchellingModel(Model):
@@ -63,6 +90,4 @@ class SchellingModel(Model):
 
         if self.happy == self.schedule.get_agent_count():
             self.running = False
-
-
